# api/api-spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Todo List REST API",
    "description": "Spesifikasi API lengkap untuk aplikasi Todo List berbasis Go, dilengkapi dengan validasi input, status code HTTP & payload, penanganan URL slug untuk tugas, serta deskripsi detail untuk tiap endpoint.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080/api/v1",
      "description": "Development Server"
    }
  ],
...
  "paths": {
...
    "/todos": {
      "get": {
        "tags": ["Todos"],
        "summary": "Mendapatkan Daftar Tugas (dengan Search & Filter)",
        "description": "Mengambil semua tugas milik pengguna yang sedang login. Mendukung pencarian kata kunci judul dan penyaringan berdasarkan ID prioritas.",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": { "type": "string", "maxLength": 150 },
            "description": "Pencarian kata kunci berdasarkan judul tugas"
          },
          {
            "name": "id_priorities",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "enum": [1, 2, 3, 4] },
            "description": "Menyaring tugas berdasarkan skala prioritas (1 - 4)"
          }
        ],
        "responses": {
          "200": {
            "description": "Daftar tugas berhasil diambil",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TodoListResponse" },
                "example": {
                  "code": 200,
                  "status": "success",
                  "message": "Berhasil mengambil daftar tugas",
                  "data": [
                    {
                      "id_todos": 1,
                      "id_users": 1,
                      "id_priorities": 1,
                      "priority": {
                        "id_priorities": 1,
                        "name": "Penting & Mendesak"
                      },
                      "title": "Mengerjakan Laporan Kinerja",
                      "slug": "mengerjakan-laporan-kinerja",
                      "description": "Harus selesai sebelum jam 5 sore",
                      "image": "http://localhost:8080/uploads/todos/task_1.jpg",
                      "is_completed": false,
                      "created_at": "2026-08-07T21:00:00Z",
                      "updated_at": "2026-08-07T21:00:00Z"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Token JWT tidak valid",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": {
                  "code": 401,
                  "status": "fail",
                  "message": "Autentikasi gagal",
                  "data": null
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Todos"],
        "summary": "Membuat Tugas Baru",
        "description": "Membuat tugas todo baru untuk pengguna yang sedang login. Slug URL dibuat otomatis dari judul. Lampiran foto bersifat opsional (max 2MB).",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": { "$ref": "#/components/schemas/TodoCreateRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Tugas baru berhasil dibuat",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TodoResponse" },
                "example": {
                  "code": 201,
                  "status": "success",
                  "message": "Tugas berhasil dibuat",
                  "data": {
                    "id_todos": 2,
                    "id_users": 1,
                    "id_priorities": 2,
                    "priority": {
                      "id_priorities": 2,
                      "name": "Penting tapi Tidak Mendesak"
                    },
                    "title": "Beli Perlengkapan Kantin",
                    "slug": "beli-perlengkapan-kantin",
                    "description": "Beli gula, kopi, dan teh",
                    "image": null,
                    "is_completed": false,
                    "created_at": "2026-08-07T21:05:00Z",
                    "updated_at": "2026-08-07T21:05:00Z"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validasi input gagal (misal judul kosong atau prioritas di luar rentang 1-4)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                },
                "example": {
                  "code": 400,
                  "status": "fail",
                  "message": "Validasi gagal",
                  "data": null,
                  "errors": {
                    "title": "Judul tugas tidak boleh kosong",
                    "id_priorities": "Prioritas harus dipilih antara skala 1 sampai 4"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Token JWT tidak valid",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": {
                  "code": 401,
                  "status": "fail",
                  "message": "Autentikasi gagal",
                  "data": null
                }
              }
            }
          }
        }
      }
    },
    "/todos/{slug}": {
      "get": {
        "tags": ["Todos"],
        "summary": "Mendapatkan Detail 1 Tugas berdasarkan Slug",
        "description": "Mengambil detail lengkap 1 tugas spesifik milik pengguna berdasarkan slug URL tugas.",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "example": "mengerjakan-laporan-kinerja",
            "description": "Slug URL unik dari tugas (contoh: `mengerjakan-laporan-kinerja`)"
          }
        ],
        "responses": {
          "200": {
            "description": "Detail tugas ditemukan",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TodoResponse" },
                "example": {
                  "code": 200,
                  "status": "success",
                  "message": "Berhasil mengambil detail tugas",
                  "data": {
                    "id_todos": 1,
                    "id_users": 1,
                    "id_priorities": 1,
                    "title": "Mengerjakan Laporan Kinerja",
                    "slug": "mengerjakan-laporan-kinerja",
                    "description": "Harus selesai sebelum jam 5 sore",
                    "image": "http://localhost:8080/uploads/todos/task_1.jpg",
                    "is_completed": false,
                    "created_at": "2026-08-07T21:00:00Z",
                    "updated_at": "2026-08-07T21:00:00Z"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Tugas tidak ditemukan atau bukan milik pengguna tersebut",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": {
                  "code": 404,
                  "status": "fail",
                  "message": "Tugas tidak ditemukan",
                  "data": null
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Todos"],
        "summary": "Mengedit Data Tugas berdasarkan Slug",
        "description": "Memperbarui data judul, deskripsi, prioritas, atau mengganti gambar pada tugas tertentu. Jika judul berubah, slug baru akan dihasilkan.",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "example": "mengerjakan-laporan-kinerja",
            "description": "Slug URL tugas yang akan diperbarui"
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": { "$ref": "#/components/schemas/TodoUpdateRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tugas berhasil diperbarui",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TodoUpdateResponse" },
                "example": {
                  "code": 200,
                  "status": "success",
                  "message": "Tugas berhasil diperbarui",
                  "data": {
                    "id_todos": 1,
                    "title": "Mengerjakan Laporan Kinerja Revisi",
                    "slug": "mengerjakan-laporan-kinerja-revisi",
                    "updated_at": "2026-08-07T21:35:00Z"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Tugas tidak ditemukan",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": {
                  "code": 404,
                  "status": "fail",
                  "message": "Tugas tidak ditemukan",
                  "data": null
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Todos"],
        "summary": "Menghapus Tugas berdasarkan Slug",
        "description": "Menghapus data tugas beserta file gambar lampirannya berdasarkan slug tugas.",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "example": "mengerjakan-laporan-kinerja",
            "description": "Slug URL tugas yang akan dihapus"
          }
        ],
        "responses": {
          "200": {
            "description": "Tugas berhasil dihapus",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DeleteResponse" },
                "example": {
                  "code": 200,
                  "status": "success",
                  "message": "Tugas berhasil dihapus"
                }
              }
            }
          },
          "404": {
            "description": "Tugas tidak ditemukan",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": {
                  "code": 404,
                  "status": "fail",
                  "message": "Tugas tidak ditemukan",
                  "data": null
                }
              }
            }
          }
        }
      }
    },
    "/todos/{slug}/toggle": {
      "patch": {
        "tags": ["Todos"],
        "summary": "Toggle Status Selesai (Undo / Redo) berdasarkan Slug",
        "description": "Mengubah status tugas dari Belum Selesai (`false`) menjadi Selesai (`true`), atau sebaliknya (Undo/Redo) menggunakan slug URL.",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "example": "mengerjakan-laporan-kinerja",
            "description": "Slug URL tugas yang akan diubah statusnya"
          }
        ],
        "responses": {
          "200": {
            "description": "Status tugas berhasil diubah",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TodoToggleResponse" },
                "example": {
                  "code": 200,
                  "status": "success",
                  "message": "Status penyelesaian tugas berhasil diperbarui",
                  "data": {
                    "id_todos": 1,
                    "slug": "mengerjakan-laporan-kinerja",
                    "is_completed": true,
                    "updated_at": "2026-08-07T21:40:00Z"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Tugas tidak ditemukan",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": {
                  "code": 404,
                  "status": "fail",
                  "message": "Tugas tidak ditemukan",
                  "data": null
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "required": ["code", "status", "message"],
        "properties": {
          "code": { "type": "integer", "description": "HTTP status code" },
          "status": { "type": "string", "enum": ["fail", "error"] },
          "message": {
            "type": "string",
            "description": "Ringkasan pesan error"
          },
          "data": {
            "type": "object",
            "nullable": true,
            "description": "Selalu null pada response error"
          }
        }
      },
      "ValidationErrorResponse": {
        "type": "object",
        "required": ["code", "status", "message", "errors"],
        "properties": {
          "code": {
            "type": "integer",
            "description": "HTTP status code (biasanya 400)"
          },
          "status": { "type": "string", "example": "fail" },
          "message": { "type": "string", "example": "Validasi gagal" },
          "data": {
            "type": "object",
            "nullable": true,
            "description": "Selalu null pada response error"
          },
          "errors": {
            "type": "object",
            "additionalProperties": { "type": "string" },
            "description": "Map pesan error per field yang gagal validasi"
          }
        }
      },
...
      "TodoPriority": {
        "type": "object",
        "description": "Prioritas yang ditampilkan di dalam objek todo",
        "properties": {
          "id_priorities": { "type": "integer", "example": 1 },
          "name": { "type": "string", "example": "Penting & Mendesak" }
        }
      },
      "Todo": {
        "type": "object",
        "properties": {
          "id_todos": { "type": "integer", "example": 1 },
          "id_users": { "type": "integer", "example": 1 },
          "id_priorities": { "type": "integer", "example": 1 },
          "priority": { "$ref": "#/components/schemas/TodoPriority" },
          "title": {
            "type": "string",
            "example": "Mengerjakan Laporan Kinerja"
          },
          "slug": {
            "type": "string",
            "example": "mengerjakan-laporan-kinerja"
          },
          "description": {
            "type": "string",
            "nullable": true,
            "example": "Harus selesai sebelum jam 5 sore"
          },
          "image": {
            "type": "string",
            "nullable": true,
            "description": "URL gambar lampiran"
          },
          "is_completed": { "type": "boolean", "default": false },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          }
        }
      },
...
      "TodoListResponse": {
        "type": "object",
        "properties": {
          "code": { "type": "integer", "example": 200 },
          "status": { "type": "string", "example": "success" },
          "message": { "type": "string" },
          "data": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Todo" }
          }
        }
      },
      "TodoResponse": {
        "type": "object",
        "properties": {
          "code": { "type": "integer" },
          "status": { "type": "string", "example": "success" },
          "message": { "type": "string" },
          "data": { "$ref": "#/components/schemas/Todo" }
        }
      },
      "TodoUpdateResponse": {
        "type": "object",
        "properties": {
          "code": { "type": "integer", "example": 200 },
          "status": { "type": "string", "example": "success" },
          "message": {
            "type": "string",
            "example": "Tugas berhasil diperbarui"
          },
          "data": {
            "type": "object",
            "properties": {
              "id_todos": { "type": "integer", "example": 1 },
              "title": { "type": "string" },
              "slug": { "type": "string" },
              "updated_at": { "type": "string", "format": "date-time" }
            }
          }
        }
      },
      "TodoToggleResponse": {
        "type": "object",
        "properties": {
          "code": { "type": "integer", "example": 200 },
          "status": { "type": "string", "example": "success" },
          "message": { "type": "string" },
          "data": {
            "type": "object",
            "properties": {
              "id_todos": { "type": "integer", "example": 1 },
              "slug": { "type": "string" },
              "is_completed": { "type": "boolean" },
              "updated_at": { "type": "string", "format": "date-time" }
            }
          }
        }
      },
      "DeleteResponse": {
        "type": "object",
        "properties": {
          "code": { "type": "integer", "example": 200 },
          "status": { "type": "string", "example": "success" },
          "message": { "type": "string", "example": "Tugas berhasil dihapus" }
        }
      },
...
      "TodoCreateRequest": {
        "type": "object",
        "required": ["title", "id_priorities"],
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 150,
            "example": "Beli Perlengkapan Kantin",
            "description": "Judul tugas (wajib, maksimal 150 karakter)"
          },
          "description": {
            "type": "string",
            "example": "Beli gula, kopi, dan teh",
            "description": "Detail catatan tugas (opsional)"
          },
          "id_priorities": {
            "type": "integer",
            "minimum": 1,
            "maximum": 4,
            "example": 2,
            "description": "ID skala prioritas dari 1 sampai 4"
          },
          "image": {
            "type": "string",
            "format": "binary",
            "description": "File gambar lampiran tugas (opsional, max 2MB)"
          }
        }
      },
      "TodoUpdateRequest": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 150,
            "description": "Judul tugas baru (opsional)"
          },
          "description": {
            "type": "string",
            "description": "Detail catatan tugas baru (opsional)"
          },
          "id_priorities": {
            "type": "integer",
            "minimum": 1,
            "maximum": 4,
            "description": "ID skala prioritas dari 1 sampai 4 (opsional)"
          },
          "image": {
            "type": "string",
            "format": "binary",
            "description": "File gambar lampiran baru (opsional, max 2MB)"
          }
        }
      }
    }
  }
}